Add Student Method Added
Database and Memory are updated after we add a new student. Also, it is checking that If you tried to add courses at the same times, the program will warn you
However, there is a problem with seeing enrolled Courses in the table view. After we restart the program, we can see that it is solved. However, we will solve that sync problem