An error fixed. Some comments added.
When we remove a course and if that course the only course a teacher gives, in that condition we could not see the teacher anymore. But now even if the teacher does not give any lessons, we can see him/her in our table, with that way we can give to him/her a new lesson.

Co-Authored-By: Yunus Emre Yalçınkaya <121396123+yunusemre009@users.noreply.github.com>